+++ src/main/java/net/parttimepolymath/spring/springkafka/services/VersionConsumer.java
@@ -1,4 +1,4 @@
-package net.parttimepolymath.spring.springkafka;
+package net.parttimepolymath.spring.springkafka.services;
 
 import lombok.extern.slf4j.Slf4j;
 import net.parttimepolymath.spring.springkafka.configuration.Version;
