+++ src/test/java/net/parttimepolymath/spring/springkafka/services/CLIParserTest.java
@@ -1,9 +1,9 @@
-package net.parttimepolymath.spring.springkafka.configuration;
+package net.parttimepolymath.spring.springkafka.services;
 
+import net.parttimepolymath.spring.springkafka.configuration.RuntimeConfig;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.DefaultApplicationArguments;
-import org.springframework.boot.test.context.SpringBootTest;
 
 import static org.junit.jupiter.api.Assertions.*;
