+++ src/test/java/net/parttimepolymath/spring/springkafka/services/StringStreamProviderTest.java
@@ -1,4 +1,4 @@
-package net.parttimepolymath.spring.springkafka;
+package net.parttimepolymath.spring.springkafka.services;
 
 import org.junit.jupiter.api.Test;
